requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-docme @ file:///home/ezzoubair/Desktop/django-docme/django-docme/dist/django-docme-0.1.tar.gz
django-mdeditor==0.1.20
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
